Catch exceptions in pointer handling
If we don't catch these exceptions, we get confused about what's going
on with the pointers, and the app basically stops working.
