+++ packages/flutter/lib/services.dart
@@ -19,5 +19,6 @@ export 'src/services/image_cache.dart';
 export 'src/services/image_decoder.dart';
 export 'src/services/image_resource.dart';
 export 'src/services/keyboard.dart';
+export 'src/services/print.dart';
 export 'src/services/service_registry.dart';
 export 'src/services/shell.dart';
