+++ packages/flutter/lib/src/gestures/pointer_router.dart
@@ -2,10 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/services.dart';
+
 import 'events.dart';
 
 /// A callback that receives a [PointerEvent]
 typedef void PointerRoute(PointerEvent event);
+
+typedef void PointerExceptionHandler(PointerRouter source, PointerEvent event, PointerRoute route, dynamic exception, StackTrace stack);
 
 /// A routing table for [PointerEvent] events.
 class PointerRouter {
@@ -34,14 +38,38 @@ class PointerRouter {
       _routeMap.remove(pointer);
   }
 
-  /// Calls the routes registed for this pointer event.
+  /// This callback is invoked whenever an exception is caught by the pointer
+  /// router. The 'source' argument is the [PointerRouter] object that caught
+  /// the exception. The 'event' argument is the pointer event that was being
+  /// routed. The 'route' argument is the callback that threw the exception. The
+  /// 'exception' argument contains the object that was thrown, and the 'stack'
+  /// argument contains the stack trace. The callback is invoked after the
+  /// information (exception, stack trace, and event; not the route callback
+  /// itself) is printed to the console.
+  PointerExceptionHandler debugPointerExceptionHandler;
+
+  /// Calls the routes registered for this pointer event.
   ///
-  /// Calls the routes in the order in which they were added to the route.
+  /// Routes are called in the order in which they were added to the
+  /// PointerRouter object.
   void route(PointerEvent event) {
     List<PointerRoute> routes = _routeMap[event.pointer];
     if (routes == null)
       return;
-    for (PointerRoute route in new List<PointerRoute>.from(routes))
-      route(event);
+    for (PointerRoute route in new List<PointerRoute>.from(routes)) {
+      try {
+        route(event);
+      } catch (exception, stack) {
+        debugPrint('-- EXCEPTION --');
+        debugPrint('The following exception was raised while routing a pointer event:');
+        debugPrint('$exception');
+        debugPrint('Stack trace:');
+        debugPrint('$stack');
+        debugPrint('Event:');
+        debugPrint('$event');
+        if (debugPointerExceptionHandler != null)
+          debugPointerExceptionHandler(this, event, route, exception, stack);
+      }
+    }
   }
 }
